Implementing distribution of blockade system
