+++ rmf_traffic/include/rmf_traffic/blockade/Rectifier.hpp
@@ -15,23 +15,24 @@
  *
 */
 
-#ifndef RMF_TRAFFIC__SCHEDULE__RECTIFIER_HPP
-#define RMF_TRAFFIC__SCHEDULE__RECTIFIER_HPP
+#ifndef RMF_TRAFFIC__BLOCKADE__RECTIFIER_HPP
+#define RMF_TRAFFIC__BLOCKADE__RECTIFIER_HPP
 
-#include <rmf_traffic/schedule/Itinerary.hpp>
-#include <rmf_traffic/schedule/ParticipantDescription.hpp>
+#include <rmf_traffic/blockade/Status.hpp>
+
+#include <rmf_utils/impl_ptr.hpp>
 
 namespace rmf_traffic {
-namespace schedule {
+namespace blockade {
 
 //==============================================================================
 /// The Rectifier class provides an interface for telling a Participant to
-/// rectify an inconsistency in the information received by a database. This
-/// rectification protocol is important when the schedule is being managed over
-/// an unreliable network.
+/// rectify an inconsistency in the information received by a moderator. This
+/// rectification protocol is important when the blockades are being managed
+/// over an unreliable network.
 ///
 /// The Rectifier class can be used by a RectifierRequester to ask a participant
-/// to retransmit a range of its past itinerary changes.
+/// to retransmit a range of its past status changes.
 ///
 /// Only the Participant class is able to create a Rectifier instance. Users of
 /// rmf_traffic cannot instantiate a Rectifier.
@@ -39,38 +40,9 @@ class Rectifier
 {
 public:
 
-  // TODO(MXG): Consider if there might be a good way to merge the functionality
-  // of this Range struct with the Inconsistency::Ranges::Range struct.
-
-  /// A range of itinerary change IDs that is currently missing from a database.
-  /// All IDs from lower to upper are missing, including lower and upper
-  /// themselves.
-  ///
-  /// It is undefined behavior if the value given to upper is less than the
-  /// value given to upper.
-  struct Range
-  {
-    /// The ID of the first itinerary change in this range that is missing
-    ItineraryVersion lower;
-
-    /// The ID of the last itinerary change in this range that is missing
-    ItineraryVersion upper;
-  };
-
-  /// Ask the participant to retransmit the specified range of its itinerary
-  /// changes.
-  ///
-  /// \param[in] ranges
-  ///   The ranges of missing Itinerary IDs
-  ///
-  /// \param[in] last_known_version
-  ///   The last ItineraryVersion known upstream.
-  void retransmit(
-    const std::vector<Range>& ranges,
-    ItineraryVersion last_known_version);
-
-  /// Get the current ItineraryVersion of the Participant.
-  ItineraryVersion current_version() const;
+  /// Check that the given status is up to date, and retransmit if any
+  /// information is out of sync.
+  void check(const Status& status);
 
   class Implementation;
 private:
@@ -90,7 +62,7 @@ private:
 ///
 /// When a schedule database reports an inconsistency for the participant tied
 /// to a RectificationRequester instance, the instance should call
-/// Rectifier::retransmit() on the Rectifier that was assigned to it.
+/// Rectifier::check() on the Rectifier that was assigned to it.
 class RectificationRequester
 {
 public:
@@ -103,7 +75,7 @@ public:
 //==============================================================================
 /// The RectificationRequesterFactory is a pure abstract interface class which
 /// should be implemented for any middlewares that intend to act as transport
-/// layers for the scheduling system.
+/// layers for the blockade system.
 class RectificationRequesterFactory
 {
 public:
@@ -112,12 +84,12 @@ public:
   ///
   /// \param[in] rectifier
   ///   This rectifier can be used by the RectificationRequester to ask the
-  ///   participant to retransmit some of its changes.
+  ///   participant to check its status.
   ///
   /// \param[in] participant_id
   ///   The ID of the participant that will hold onto this
   ///   RectificationRequester. This is the same participant that the rectifier
-  ///   will request retransmissions to.
+  ///   will request for checks.
   virtual std::unique_ptr<RectificationRequester> make(
     Rectifier rectifier,
     ParticipantId participant_id) = 0;
@@ -128,33 +100,32 @@ public:
 };
 
 //==============================================================================
-// Forward declaration for DatabaseRectificationRequesterFactory
-class Database;
+// Forward declaration for ModeratorRectificationRequesterFactory
+class Moderator;
 
 //==============================================================================
 /// This class provides a simple implementation of a
-/// RectificationRequesterFactory that just hooks directly into a Database
+/// RectificationRequesterFactory that just hooks directly into a Moderator
 /// instance and issues rectification requests when told to based on the current
 /// inconsistencies in the Database.
-class DatabaseRectificationRequesterFactory
-  : public RectificationRequesterFactory
+class ModeratorRectificationRequesterFactory
+    : public RectificationRequesterFactory
 {
 public:
 
-  /// This accepts a const-reference to a Database instance. Note that this
-  /// class will store a reference to this Database, so its lifecycle is
-  /// implicitly dependent on the Database's lifecycle.
-  //
-  // TODO(MXG): Should this demand a std::shared_ptr<Database> instead?
-  DatabaseRectificationRequesterFactory(const Database& database);
+  /// Constructor
+  ///
+  /// \param[in] moderator
+  ///   The moderator object that this will rectify for.
+  ModeratorRectificationRequesterFactory(std::shared_ptr<Moderator> moderator);
 
   // Documentation inherited
   std::unique_ptr<RectificationRequester> make(
-    Rectifier rectifier,
-    ParticipantId participant_id) final;
+      Rectifier rectifier,
+      ParticipantId participant_id) final;
 
-  /// Call this function to instruct all the RectificationRequestors produced
-  /// by this factory to perform their rectifications.
+  /// Call this function to instruct all the RectificationRequesters produced by
+  /// this factory to perform their rectifications.
   void rectify();
 
   class Implementation;
@@ -162,7 +133,8 @@ private:
   rmf_utils::unique_impl_ptr<Implementation> _pimpl;
 };
 
-} // namespace schedule
+
+} // namespace blockade
 } // namespace rmf_traffic
 
-#endif // RMF_TRAFFIC__SCHEDULE__RECTIFIER_HPP
+#endif // RMF_TRAFFIC__BLOCKADE__RECTIFIER_HPP
